New approach to splitting
Not splitting with an initial depth (and the planned further
splitting in a second step). Instead starting with only one inital
rectangle and filling it up to a maximum number of nodes. Once that
limit is reached, split the rectangle, create a new output for each of
the children, split the file created before by distributing the nodes
to both of the children's files. Then delete the file and continue
distributing nodes from input.