First version of various node array file types
A node array is a file that stores all nodes from an OSM file in an
array. Only coordinates are stored, no tags, no metadata. It is intended
for resolving the geometry of ways and relations quickly.

Each node is stored with fixed size. Missing intermediate nodes are left
blank and filled with a null value, if supported by the implementation.
There are currently four implementations offering varying precision and
storage properties:

* NodeArrayDouble: 16 bytes per node, 8 bytes for each coordinate, full
  double value range.
* NodeArrayFloat: 8 bytes per node, 4 bytes for each coordinate, full
  float value range.
* NodeArrayInteger: 8 bytes per node, 4 bytes for each coordinate,
  encoded as integers.
* NodeArrayShort: 4 bytes per node, 2 bytes for each coordinate,
  encoded as integers.