New version of CreateNodeTreeMaxNodes
Using a two-pass approach to build the tree. In a the first pass, we
create a big tree with a lot of leafs by splitting every node up to a
certain depth. Then we count the number of data points that map to each
leaf by going through the input data a first time. After that we melt
tree nodes bottom up to eliminate as many nodes as possible without
violating the max_nodes parameter.

After that we iterate the data again and perform the actual distribution
of data to data files for each leaf.

The tree is initially split up to a limited depth. Hence, it can happen
that we end up with leafs that have too many data nodes. Thus we run the
same two-pass algorithm again for nodes that violate the max_nodes
parameter. Depending on the parameters this can happen after the second
iteration as well, so we apply this method recursively until the limits
are achieved.