Tools for extracting entities using id lists
Added Extractor class that will iterate an osm data file thereby
distributing entities to a number of output files. It uses a set of
IdListInputStreams to determine which entities to select from the input
stream.

Added the ExtractEntities executable that uses the Extractor to extract
from one input source to separate output files the elements that are
references by a set of id list files.