Relation queries: fixed bug
Not adding the ids of nodes, ways and waynodes of relations found into
the respective id sets that are used to retrieve further relations based
on the fact that any of their members are contained withing these sets.
This should prevent some relations to end up in the query results that
do not actually intersect the specified query area.