Query classes support relation selection
The Query class and its subclasses support an optional (nullable)
RelationFilter argument that can be used to select a subset of relations
when querying relation batches. Using this approach a query can reduce
the resulting filesize by omitting relations that are irrelevant for the
use case of the extraction. Since some relation networks tend to be huge,
this can reduce the filesize drastically depending on the filtered out
relations.